glibc: Move `signal.h` definitions to `src/new`
This allows for quite a bit of cleanup. Links:

* https://github.com/sailfishos-mirror/glibc/blob/ae646973c5957b7eed06cb80d49d13b42178072d/signal/signal.h
* https://github.com/sailfishos-mirror/glibc/blob/ae646973c5957b7eed06cb80d49d13b42178072d/sysdeps/unix/sysv/linux/bits/sigaction.h
* https://github.com/sailfishos-mirror/glibc/blob/ae646973c5957b7eed06cb80d49d13b42178072d/sysdeps/unix/sysv/linux/mips/bits/sigaction.h
* https://github.com/sailfishos-mirror/glibc/blob/ae646973c5957b7eed06cb80d49d13b42178072d/sysdeps/unix/sysv/linux/s390/bits/sigaction.h
* https://github.com/sailfishos-mirror/glibc/blob/ae646973c5957b7eed06cb80d49d13b42178072d/sysdeps/unix/sysv/linux/sparc/bits/sigaction.h
* https://github.com/sailfishos-mirror/glibc/blob/ae646973c5957b7eed06cb80d49d13b42178072d/bits/signum-generic.h#L76
* https://github.com/sailfishos-mirror/glibc/blob/ae646973c5957b7eed06cb80d49d13b42178072d/sysdeps/unix/sysv/linux/bits/signum-arch.h
* https://github.com/sailfishos-mirror/glibc/blob/ae646973c5957b7eed06cb80d49d13b42178072d/sysdeps/unix/sysv/linux/mips/bits/signum-arch.h
* https://github.com/sailfishos-mirror/glibc/blob/ae646973c5957b7eed06cb80d49d13b42178072d/sysdeps/unix/sysv/linux/sparc/bits/signum-arch.h

(backport <https://github.com/rust-lang/libc/pull/5408>)
(cherry picked from commit 72f013506e8e58e4e4f8d3c506c2028037db6543)
