+++ src/unix/linux_like/linux/gnu/b32/m68k/mod.rs
@@ -7,15 +7,6 @@ use crate::{
 pub type wchar_t = i32;
 
 s! {
-    // FIXME(1.0): This should not implement `PartialEq`
-    #[allow(unpredictable_function_pointer_comparisons)]
-    pub struct sigaction {
-        pub sa_sigaction: crate::sighandler_t,
-        pub sa_mask: crate::sigset_t,
-        pub sa_flags: c_int,
-        pub sa_restorer: Option<extern "C" fn()>,
-    }
-
     pub struct statfs {
         pub f_type: crate::__fsword_t,
         pub f_bsize: crate::__fsword_t,
@@ -289,9 +280,6 @@ pub const ENOTRECOVERABLE: c_int = 131;
 pub const EHWPOISON: c_int = 133;
 pub const ERFKILL: c_int = 132;
 
-pub const SA_SIGINFO: c_int = 0x00000004;
-pub const SA_NOCLDWAIT: c_int = 0x00000002;
-
 pub const SOCK_STREAM: c_int = 1;
 pub const SOCK_DGRAM: c_int = 2;
 
@@ -314,29 +302,6 @@ pub const POLLWRBAND: c_short = 0x200;
 pub const EFD_NONBLOCK: c_int = 0x800;
 pub const SFD_NONBLOCK: c_int = 0x0800;
 
-pub const SIGCHLD: c_int = 17;
-pub const SIGBUS: c_int = 7;
-pub const SIGUSR1: c_int = 10;
-pub const SIGUSR2: c_int = 12;
-pub const SIGCONT: c_int = 18;
-pub const SIGSTOP: c_int = 19;
-pub const SIGTSTP: c_int = 20;
-pub const SIGURG: c_int = 23;
-pub const SIGIO: c_int = 29;
-pub const SIGSYS: c_int = 31;
-pub const SIGSTKFLT: c_int = 16;
-pub const SIGPOLL: c_int = 29;
-pub const SIGPWR: c_int = 30;
-pub const SIG_SETMASK: c_int = 2;
-pub const SIG_BLOCK: c_int = 0x000000;
-pub const SIG_UNBLOCK: c_int = 0x01;
-pub const SIGTTIN: c_int = 21;
-pub const SIGTTOU: c_int = 22;
-pub const SIGXCPU: c_int = 24;
-pub const SIGXFSZ: c_int = 25;
-pub const SIGVTALRM: c_int = 26;
-pub const SIGPROF: c_int = 27;
-pub const SIGWINCH: c_int = 28;
 pub const SIGSTKSZ: size_t = 8192;
 pub const MINSIGSTKSZ: size_t = 2048;
 pub const CBAUD: crate::tcflag_t = 0o0010017;
