+++ src/unix/linux_like/linux/gnu/b32/mips/mod.rs
@@ -166,16 +166,6 @@ s! {
         __f_spare: [c_int; 6],
     }
 
-    // FIXME(1.0): This should not implement `PartialEq`
-    #[allow(unpredictable_function_pointer_comparisons)]
-    pub struct sigaction {
-        pub sa_flags: c_int,
-        pub sa_sigaction: crate::sighandler_t,
-        pub sa_mask: crate::sigset_t,
-        pub sa_restorer: Option<extern "C" fn()>,
-        _resv: [c_int; 1],
-    }
-
     pub struct stack_t {
         pub ss_sp: *mut c_void,
         pub ss_size: size_t,
@@ -764,33 +754,6 @@ pub const MAP_STACK: c_int = 0x40000;
 
 pub const SOCK_STREAM: c_int = 2;
 pub const SOCK_DGRAM: c_int = 1;
-
-pub const SA_SIGINFO: c_int = 0x00000008;
-pub const SA_NOCLDWAIT: c_int = 0x00010000;
-
-pub const SIGEMT: c_int = 7;
-pub const SIGCHLD: c_int = 18;
-pub const SIGBUS: c_int = 10;
-pub const SIGTTIN: c_int = 26;
-pub const SIGTTOU: c_int = 27;
-pub const SIGXCPU: c_int = 30;
-pub const SIGXFSZ: c_int = 31;
-pub const SIGVTALRM: c_int = 28;
-pub const SIGPROF: c_int = 29;
-pub const SIGWINCH: c_int = 20;
-pub const SIGUSR1: c_int = 16;
-pub const SIGUSR2: c_int = 17;
-pub const SIGCONT: c_int = 25;
-pub const SIGSTOP: c_int = 23;
-pub const SIGTSTP: c_int = 24;
-pub const SIGURG: c_int = 21;
-pub const SIGIO: c_int = 22;
-pub const SIGSYS: c_int = 12;
-pub const SIGPOLL: c_int = 22;
-pub const SIGPWR: c_int = 19;
-pub const SIG_SETMASK: c_int = 3;
-pub const SIG_BLOCK: c_int = 0x1;
-pub const SIG_UNBLOCK: c_int = 0x2;
 
 pub const POLLWRNORM: c_short = 0x004;
 pub const POLLWRBAND: c_short = 0x100;
