+++ src/unix/linux_like/linux/gnu/b32/sparc/mod.rs
@@ -9,15 +9,6 @@ use crate::{
 pub type wchar_t = i32;
 
 s! {
-    // FIXME(1.0): This should not implement `PartialEq`
-    #[allow(unpredictable_function_pointer_comparisons)]
-    pub struct sigaction {
-        pub sa_sigaction: crate::sighandler_t,
-        pub sa_mask: crate::sigset_t,
-        pub sa_flags: c_int,
-        pub sa_restorer: Option<extern "C" fn()>,
-    }
-
     pub struct statfs {
         pub f_type: crate::__fsword_t,
         pub f_bsize: crate::__fsword_t,
@@ -315,33 +306,6 @@ pub const ERFKILL: c_int = 134;
 
 pub const SOCK_STREAM: c_int = 1;
 pub const SOCK_DGRAM: c_int = 2;
-
-pub const SA_SIGINFO: c_int = 0x200;
-pub const SA_NOCLDWAIT: c_int = 0x100;
-
-pub const SIGEMT: c_int = 7;
-pub const SIGTTIN: c_int = 21;
-pub const SIGTTOU: c_int = 22;
-pub const SIGXCPU: c_int = 24;
-pub const SIGXFSZ: c_int = 25;
-pub const SIGVTALRM: c_int = 26;
-pub const SIGPROF: c_int = 27;
-pub const SIGWINCH: c_int = 28;
-pub const SIGCHLD: c_int = 20;
-pub const SIGBUS: c_int = 10;
-pub const SIGUSR1: c_int = 30;
-pub const SIGUSR2: c_int = 31;
-pub const SIGCONT: c_int = 19;
-pub const SIGSTOP: c_int = 17;
-pub const SIGTSTP: c_int = 18;
-pub const SIGURG: c_int = 16;
-pub const SIGIO: c_int = 23;
-pub const SIGSYS: c_int = 12;
-pub const SIGPOLL: c_int = 23;
-pub const SIGPWR: c_int = 29;
-pub const SIG_SETMASK: c_int = 4;
-pub const SIG_BLOCK: c_int = 1;
-pub const SIG_UNBLOCK: c_int = 2;
 
 pub const POLLWRNORM: c_short = 4;
 pub const POLLWRBAND: c_short = 0x100;
